#87: Update dao-mongo module to use the v3 API

+++ core/dao/src/main/java/uk/gov/dwp/queue/triage/core/dao/FailedMessageDao.java
@@ -21,7 +21,7 @@ public interface FailedMessageDao {
 
     List<StatusHistoryEvent> getStatusHistory(FailedMessageId failedMessageId);
 
-    int removeFailedMessages();
+    long removeFailedMessages();
 
     void addLabel(FailedMessageId failedMessageId, String label);
